Remove scheduled issue aos task and refactor
