+++ src/integrationTest/resources/application-schedule_test.yaml
@@ -1,3 +1,2 @@
 schedule:
-  issue_aos: '*/1 * * * * *'
   awaiting_conditional_order: '*/1 * * * * *'
