+++ src/integrationTest/resources/application.yaml
@@ -81,7 +81,6 @@ send-letter:
   url: ${SEND_LETTER_SERVICE_BASEURL:http://rpe-send-letter-service-aat.service.core-compute-aat.internal}
 
 schedule:
-  issue_aos: '0 0 0 * * ?' # Midnight each day
   awaiting_conditional_order: '0 0 0 * * ?' # Midnight each day
 
 aos_pack:
