+++ src/main/resources/application.yaml
@@ -130,7 +130,6 @@ send-letter:
   url: ${SEND_LETTER_SERVICE_BASEURL:http://rpe-send-letter-service-aat.service.core-compute-aat.internal}
 
 schedule:
-  issue_aos: '0 0 0 * * ?' # Midnight each day
   awaiting_conditional_order: ${AWAITING_CONDITIONAL_ORDER_SCHEDULED_CRON:0 0 0 * * ?}
 
 aos_pack:
